# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is simple REST API for a UK train data.",
        "title": "Train API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.train.vulpz.com",
    "basePath": "/",
    "paths": {
...
        "/station/{stationId}": {
            "get": {
                "description": "Responds with a single station in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stations"
                ],
                "summary": "Get Station",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Station Id (TIPLOC)",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Station"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stations": {
            "get": {
                "description": "Responds with a list of all stations in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stations"
                ],
                "summary": "List Stations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Station"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stations.geojson": {
            "get": {
                "description": "Responds with a GeoJSON representing station locations as features",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stations"
                ],
                "summary": "Get Station GeoJSON",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GeoJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "GeoJSON": {
            "type": "object",
            "properties": {
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GeoJSON"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "Station": {
            "type": "object",
            "properties": {
                "crs": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nlc": {
                    "type": "string"
                },
                "tiploc": {
                    "type": "string"
                }
            }
        }
    }
}